refactor : Entity -> DTO로 변환하는 과정에서 Common 모듈이 Core 모듈을 의존하게 되는 상황이 발생. 결론적으로 DTO는 API 모듈에서 관리하도록 리팩토링 진행.
