+++ credit-api/src/main/java/com/api/dto/response/ContractAgreeResponse.java
@@ -1,17 +1,16 @@
-package com.credit.common.contract.response;
+package com.api.dto.response;
 
-import com.credit.common.contract.ContractStatus;
+import com.creditcore.enums.contract.ContractStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @Builder
-public class RecoveryProgramCreateResponse {
-
+public class ContractAgreeResponse {
     private String contractId;
     private ContractStatus status;
 }
