+++ credit-api/src/main/java/com/api/dto/response/RecoveryProgramAgreeResponse.java
@@ -1,16 +1,16 @@
-package com.credit.common.contract.response;
+package com.api.dto.response;
 
-import com.credit.common.contract.ContractStatus;
+import com.creditcore.enums.contract.ContractStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @Builder
-public class ContractAgreeResponse {
+public class RecoveryProgramAgreeResponse {
     private String contractId;
     private ContractStatus status;
 }
