+++ credit-api/src/main/java/com/api/dto/response/RecoveryProgramCreateResponse.java
@@ -1,6 +1,6 @@
-package com.credit.common.contract.response;
+package com.api.dto.response;
 
-import com.credit.common.contract.ContractStatus;
+import com.creditcore.enums.contract.ContractStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,7 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class RecoveryProgramAgreeResponse {
+public class RecoveryProgramCreateResponse {
+
     private String contractId;
     private ContractStatus status;
 }
